All service models were separated to directories by service name
