+++ VacationRental.Api.Tests.Unit/Services/CalendarServiceTests.cs
@@ -11,6 +11,7 @@ using VacationRental.Api.Repositories;
 using VacationRental.Api.Services;
 using VacationRental.Api.Services.Implementation;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Calendar;
 using VacationRental.Api.Tests.Unit.DSL;
 using VacationRental.Api.Tests.Unit.Extensions;
