+++ VacationRental.Api.Tests.Unit/Services/RentalServiceTests.cs
@@ -8,6 +8,7 @@ using VacationRental.Api.Repositories;
 using VacationRental.Api.Services;
 using VacationRental.Api.Services.Implementation;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Rental;
 using VacationRental.Api.Tests.Unit.DSL;
 using VacationRental.Api.Tests.Unit.Extensions;
