+++ VacationRental.Api/Controllers/v1/BookingsController.cs
@@ -7,6 +7,7 @@ using VacationRental.Api.Contracts.v1.Common;
 using VacationRental.Api.Controllers.v1.Mappers;
 using VacationRental.Api.Services;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Booking;
 
 namespace VacationRental.Api.Controllers.v1
 {
