+++ VacationRental.Api/Controllers/v1/RentalsController.cs
@@ -7,6 +7,7 @@ using VacationRental.Api.Contracts.v1.Rental;
 using VacationRental.Api.Controllers.v1.Mappers;
 using VacationRental.Api.Services;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Rental;
 
 namespace VacationRental.Api.Controllers.v1
 {
