+++ VacationRental.Api/Services/ICalendarService.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading;
 using System.Threading.Tasks;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Calendar;
 
 namespace VacationRental.Api.Services;
