+++ VacationRental.Api/Services/IRentalService.cs
@@ -2,6 +2,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using VacationRental.Api.Models;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Rental;
 
 namespace VacationRental.Api.Services;
