+++ VacationRental.Api/Services/Implementation/CalendarService.cs
@@ -7,6 +7,7 @@ using VacationRental.Api.Logging.Extensions.Services;
 using VacationRental.Api.Models;
 using VacationRental.Api.Repositories;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Calendar;
 
 namespace VacationRental.Api.Services.Implementation;
