+++ VacationRental.Api/Services/Implementation/RentalService.cs
@@ -5,6 +5,7 @@ using System.Threading.Tasks;
 using VacationRental.Api.Models;
 using VacationRental.Api.Repositories;
 using VacationRental.Api.Services.Models;
+using VacationRental.Api.Services.Models.Rental;
 
 namespace VacationRental.Api.Services.Implementation;
