+++ VacationRental.Api/Services/Models/Booking/CreateBookingResultErrorStatus.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics.CodeAnalysis;
 
-namespace VacationRental.Api.Services.Models;
+namespace VacationRental.Api.Services.Models.Booking;
 
 [SuppressMessage("ReSharper", "UnusedMember.Global")]
 public enum CreateBookingResultErrorStatus
