+++ VacationRental.Api/Services/Models/Rental/CreateRentalResult.cs
@@ -1,11 +1,10 @@
 using System.Diagnostics.CodeAnalysis;
-using VacationRental.Api.Models;
 
-namespace VacationRental.Api.Services.Models;
+namespace VacationRental.Api.Services.Models.Rental;
 
 public class CreateRentalResult
 {
-    private CreateRentalResult(bool isSuccess, CreateRentalResultErrorStatus errorStatus, string? errorMessage, Rental? rental)
+    private CreateRentalResult(bool isSuccess, CreateRentalResultErrorStatus errorStatus, string? errorMessage, Api.Models.Rental? rental)
     {
         IsSuccess = isSuccess;
         ErrorStatus = errorStatus;
@@ -18,10 +17,10 @@ public class CreateRentalResult
     public bool IsSuccess { get; }
     public CreateRentalResultErrorStatus ErrorStatus { get; }
     public string? ErrorMessage { get; }
-    public Rental? Rental { get; }
+    public Api.Models.Rental? Rental { get; }
 
     public static CreateRentalResult ValidationFail(string message) 
         => new(false, CreateRentalResultErrorStatus.ValidationFail, message, null);
-    public static CreateRentalResult Successful(Rental rental) 
+    public static CreateRentalResult Successful(Api.Models.Rental rental) 
         => new(true, CreateRentalResultErrorStatus.Undefined, null, rental);
 }
